const 부분 분리 시키고 eume형태로 로직 수정중
