+++ lib/common/audio_file_path.dart
@@ -1,0 +1,8 @@
+//장구 장단
+const String JOONGJOONG = 'assets/music/joongjoong.mp3';
+const String JAJIN = 'assets/music/jajin.mp3';
+const String HUIMORI = 'assets/music/huimori.mp3';
+const String GOOD = 'assets/music/good.mp3';
+const String SEMACHI = 'assets/music/semachi.mp3';
+// 배경음
+const String backMusic = 'assets/music/danso_arirang_mixed.mp3';
