+++ lib/common/icon.dart
@@ -1,0 +1,102 @@
+//에셋(assets) 가져오기
+const List<String> DANSO_SVG_LIST = [
+  'assets/images/danso_code/ic_jung.svg',
+  'assets/images/danso_code/ic_lim.svg',
+  'assets/images/danso_code/ic_mu.svg',
+  'assets/images/danso_code/ic_hwang.svg',
+  'assets/images/danso_code/ic_tea.svg',
+  'assets/images/danso_code/ic_jung.svg',
+  'assets/images/danso_code/ic_lim.svg',
+  'assets/images/danso_code/ic_mu.svg',
+  'assets/images/danso_code/ic_hwang.svg',
+  'assets/images/danso_code/ic_ high_tea.svg',
+];
+const List<String> DANSO_CODE_SVG_LIST = [
+  'assets/images/danso_code/ic_jung.svg',
+  'assets/images/danso_code/ic_lim.svg',
+  'assets/images/danso_code/ic_mu.svg',
+  'assets/images/danso_code/ic_hwang.svg',
+  'assets/images/danso_code/ic_tea.svg',
+  'assets/images/danso_code/ic_ high_tea.svg',
+];
+//업 다운 svg
+const String UP_SVG = 'assets/images/ic_up.svg';
+const String DOWN_SVG = 'assets/images/ic_down.svg';
+//운지법
+
+//main svg 기준음잡디 메인페이지 아이콘
+const String HOME_SVG = 'assets/images/ic_home.svg';
+const String BOLW_SVG = 'assets/images/ic_blow.svg';
+const String LOGO_SVG = 'assets/images/ic_logo.svg';
+const String SUCCESS_SVG = 'assets/images/ic_success.svg';
+const String FAIL_SVG = 'assets/images/ic_failure.svg';
+const String WARNING_SVG = 'assets/images/ic_warning.svg';
+const String MAIN_ILL1_SVG =
+    'assets/images/mainImage/main_illeoseuteu_music(1).svg';
+const String MAIN_ILL2_SVG =
+    'assets/images/mainImage/main_illeoseuteu_music(2).svg';
+const String ON_SVG = 'assets/images/mainImage/ic_on.svg';
+const String OFF_SVG = 'assets/images/mainImage/ic_off.svg';
+const String Q_AND_A_SVG = 'assets/images/q&a_sp.svg';
+const String TIP_SVG = 'assets/images/tip_sp.svg';
+const String DANSO_LESSON_SVG = 'assets/images/ic_danso_lesson.svg';
+const String DANSO_STUDY_SVG = 'assets/images/ic_danso_study.svg';
+const String NOTO_BOLD = 'assets/font/NotoSansKR-Bold.otf';
+const String NOTO_MEDIUM = 'assets/font/NotoSansKR-Medium.otf';
+const String NOTO_REGULAR = 'assets/font/NotoSansKR-Regular.otf';
+//마이 페이지 아이콘
+const String BOOKMARK_SVG = 'assets/images/my_page/ic_bookmark.svg';
+const String GRAPH_SVG = 'assets/images/my_page/ic_graph.svg';
+const String MP3_SVG = 'assets/images/my_page/ic_mp3.svg';
+const String MY_MUSIC_SVG = 'assets/images/my_page/ic_my_music.svg';
+const String PLAY_STOP_SVG = 'assets/images/my_page/ic_play_stop.svg';
+const String PLAY_SVG = 'assets/images/my_page/ic_play.svg';
+const String MY_RECORD_SVG = 'assets/images/my_page/ic_my_record.svg';
+const String SEE_MORE_SVG = 'assets/images/my_page/ic_see_more.svg';
+const String VIDEO_SVG = 'assets/images/my_page/ic_video.svg';
+//녹음 녹화 아이콘
+const String RECORD_SVG = 'assets/images/ic_record.svg';
+const String TRANSCRIPTION_SVG = 'assets/images/ic_transcription.svg';
+
+//연주곡 익히기부분
+const List STEP_SVG = [
+  '',
+  'assets/images/step/ic_step_1.svg',
+  'assets/images/step/ic_step_2.svg',
+  'assets/images/step/ic_step_3.svg',
+  'assets/images/step/ic_step_4.svg',
+];
+const List COUNT_LIST_SVG = [
+  'assets/images/step/ic_music_list_1.svg',
+  'assets/images/step/ic_music_list_2.svg',
+  'assets/images/step/ic_music_list_3.svg',
+  'assets/images/step/ic_music_list_4.svg',
+  'assets/images/step/ic_music_list_5.svg',
+  'assets/images/step/ic_music_list_6.svg',
+  'assets/images/step/ic_music_list_7.svg',
+];
+const String LEFT_SVG = 'assets/images/step/ic_left.svg';
+const String RIGHT_SVG = 'assets/images/step/ic_right.svg';
+//노래끝난후 축하부분
+const List STAR_SVG = [
+  'assets/images/star/ic_star_1.svg',
+  'assets/images/star/ic_star_2.svg',
+  'assets/images/star/ic_star_3.svg',
+];
+//메인 아이콘
+
+const String DANSO_TUNING_SVG = 'assets/images/mainImage/ic_danso_tuning.svg';
+const String INFOR_SVG = 'assets/images/mainImage/ic_infor.svg';
+const String QandA_SVG = 'assets/images/mainImage/ic_q&a.svg';
+const String STUDY_SVG = 'assets/images/mainImage/ic_study.svg';
+const String TUNE_SVG = 'assets/images/mainImage/ic_tune.svg';
+
+//티어별 svg
+const List TEAR_SVG_LIST = [
+  'assets/images/tear/ic_unranked.svg',
+  'assets/images/tear/ic_bronze.svg',
+  'assets/images/tear/ic_silver.svg',
+  'assets/images/tear/ic_gold.svg',
+  'assets/images/tear/ic_platinum.svg',
+  'assets/images/tear/ic_master.svg',
+];
