+++ lib/common/youtude.dart
@@ -1,0 +1,45 @@
+const String DANSO_HISTORY_URL = 'j_XrCCOwBe8';
+const String DANSO_KIND_URL = 'AvH2NGdYLjk';
+const String DANSO_PALY_URL = 'hogSau50f64';
+const String DANSO_UNM_URL = 'rg5oT4VLA14';
+const String DANSO_SAME_UNM_TITLE_URL = 'o8MulHBatrM';
+const String DANSO_UNM_TEST_URL = 'C-utyYK-FDI';
+//아래쪽은 선생님들url
+const String DANSO_LIPS_URL = 'Cuc3e0v1Dc0';
+const String DANSO_LIPS2_URL = 'aoqaQKH3iZg';
+const String DANSO_BREATH_URL = 'opQtzo9zXJE';
+const String DANSO_BREATH2_URL = 'k-PYlJKY8Uc';
+
+const List Q_AND_A_URL = [
+  'JHVGwgMXfAo',
+  'PBAZnJNcYVY',
+  'PBAZnJNcYVY',
+  'PBAZnJNcYVY',
+  'PBAZnJNcYVY',
+  'PBAZnJNcYVY',
+  'PBAZnJNcYVY',
+  'PBAZnJNcYVY',
+];
+
+const List TIP_URL = [
+  'GiFWeblNJc8',
+  'TDprjZyakoQ',
+  'z2afsvDPhd8',
+  '4U_cOC6f66A',
+  'EghU5mA8KTw',
+];
+const List TIP_TIME = [0.0, 0.0, 0.0, 0.0];
+const List Q_AND_A_TIME = [
+  0.0,
+  0.0,
+  38.0,
+  211.0,
+  308.0,
+  351.0,
+  378.0,
+  527.0,
+  616.0
+];
+//youtube Size style
+const double youtubeWidth = 360;
+const double youtubeHeight = 250;
