+++ lib/controllers/my_page_controller.dart
@@ -1,7 +1,8 @@
 import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:get/get.dart';
+import 'package:project_danso/common/audio_file_path.dart';
 
-import 'package:project_danso/common/const.dart';
+import 'package:project_danso/common/contant.dart';
 
 class MyPageController extends GetxController {
   bool starStopState = false;
