+++ lib/screens/main_danso_learning_level_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:project_danso/common/const.dart';
+import 'package:project_danso/common/contant.dart';
 import 'package:project_danso/widgets/widgets.dart';
 
 class MainDansoLearningLevelScreen extends StatelessWidget {
