+++ lib/widgets/base_widgets/default_appbar.dart
@@ -1,15 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:project_danso/common/color.dart';
-import 'package:project_danso/common/const.dart';
+import 'package:project_danso/common/contant.dart';
+import 'package:project_danso/common/size.dart';
 
 PreferredSizeWidget defaultAppBar({required String title}) {
   return AppBar(
     title: Text(
       '$title',
-      style: TextStyle(color: white, fontSize: textBasicSize.sp),
+      style: TextStyle(
+          color: MctColor.white.getMctColor,
+          fontSize: MctSize.fourteen.getSize.sp),
     ),
     elevation: 1,
-    backgroundColor: getMctColor(MctColor.indiDarkGray),
+    backgroundColor: MctColor.indiDarkGray.getMctColor,
   );
 }
