+++ lib/widgets/base_widgets/picture_and_text.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:project_danso/common/const.dart';
+import 'package:project_danso/common/contant.dart';
+import 'package:project_danso/common/size.dart';
 
 class PictureAndText extends StatelessWidget {
   String title;
@@ -31,7 +32,7 @@ class PictureAndText extends StatelessWidget {
           ),
           SizedBox(height: 20.h),
           Container(
-            margin: EdgeInsets.symmetric(horizontal: basicPadding),
+            margin: EdgeInsets.symmetric(horizontal: MctSize.fifteen.getSize),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
@@ -43,7 +44,8 @@ class PictureAndText extends StatelessWidget {
                   ),
                 ),
                 SizedBox(height: 10.h),
-                Text(contant, style: TextStyle(fontSize: textBasicSize.sp)),
+                Text(contant,
+                    style: TextStyle(fontSize: MctSize.fourteen.getSize.sp)),
               ],
             ),
           ),
