+++ lib/widgets/camera_record/song_camara_recoding.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:project_danso/common/const.dart';
+import 'package:project_danso/common/color.dart';
+import 'package:project_danso/common/contant.dart';
+import 'package:project_danso/common/size.dart';
 import 'package:project_danso/controllers/camera_record/camera_record_controller.dart';
 import 'package:project_danso/controllers/controllers.dart';
 import 'package:camera/camera.dart';
@@ -79,14 +81,16 @@ class _SongCamaraRecodingState extends State<SongCamaraRecoding> {
               height: 30.h,
               child: ElevatedButton(
                 child: Text(caController.recordingText,
-                    style: TextStyle(fontSize: textSmallSize.sp)),
+                    style: TextStyle(fontSize: MctSize.twelve.getSize.sp)),
                 style: ElevatedButton.styleFrom(
                     elevation: 0,
-                    primary: white,
-                    onPrimary: buttonColorOrang,
-                    side: BorderSide(color: buttonColorOrang),
-                    textStyle:
-                        TextStyle(fontSize: 12.sp, color: buttonColorOrang)),
+                    primary: MctColor.white.getMctColor,
+                    onPrimary: MctColor.buttonColorOrange.getMctColor,
+                    side: BorderSide(
+                        color: MctColor.buttonColorOrange.getMctColor),
+                    textStyle: TextStyle(
+                        fontSize: 12.sp,
+                        color: MctColor.buttonColorOrange.getMctColor)),
                 onPressed: () async {
                   caController.isRecordingState();
 
@@ -94,7 +98,8 @@ class _SongCamaraRecodingState extends State<SongCamaraRecoding> {
                   if (caController.isRecording) {
                     await Get.dialog(
                       Dialog(
-                          backgroundColor: Colors.white.withOpacity(0),
+                          backgroundColor:
+                              MctColor.white.getMctColor.withOpacity(0),
                           elevation: 0,
                           child: GameTimerWidget()),
                       barrierDismissible: false,
@@ -122,15 +127,17 @@ class _SongCamaraRecodingState extends State<SongCamaraRecoding> {
               width: 78.w,
               height: 30.h,
               child: ElevatedButton(
-                child:
-                    Text('반주만', style: TextStyle(fontSize: textSmallSize.sp)),
+                child: Text('반주만',
+                    style: TextStyle(fontSize: MctSize.twelve.getSize.sp)),
                 style: ElevatedButton.styleFrom(
                     elevation: 0,
-                    primary: white,
-                    onPrimary: buttonColorOrang,
-                    side: BorderSide(color: buttonColorOrang),
-                    textStyle:
-                        TextStyle(fontSize: 12.sp, color: buttonColorOrang)),
+                    primary: MctColor.white.getMctColor,
+                    onPrimary: MctColor.buttonColorOrange.getMctColor,
+                    side: BorderSide(
+                        color: MctColor.buttonColorOrange.getMctColor),
+                    textStyle: TextStyle(
+                        fontSize: 12.sp,
+                        color: MctColor.buttonColorOrange.getMctColor)),
                 onPressed: () {},
               ),
             ),
@@ -139,7 +146,7 @@ class _SongCamaraRecodingState extends State<SongCamaraRecoding> {
         SizedBox(height: 3.h),
         Text(
           widget.jandan,
-          style: TextStyle(fontSize: textSmallSize.sp),
+          style: TextStyle(fontSize: MctSize.twelve.getSize.sp),
         ),
       ],
     );
