+++ lib/widgets/my_page/my_page_look.dart
@@ -6,7 +6,9 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:project_danso/common/color.dart';
-import 'package:project_danso/common/const.dart';
+import 'package:project_danso/common/contant.dart';
+import 'package:project_danso/common/icon.dart';
+import 'package:project_danso/common/size.dart';
 import 'package:project_danso/controllers/audio_and_video_list_controller.dart';
 import 'package:project_danso/controllers/audio_and_video_db_controller.dart';
 import 'package:project_danso/controllers/controllers.dart';
@@ -38,7 +40,7 @@ class _MyPageLookState extends State<MyPageLook> {
       appBar: tabbarAndAppBar(title: '연주보기', tabbar: null, enableTabBar: false),
       body: Obx(
         () => ListView.builder(
-          padding: const EdgeInsets.all(basicPadding),
+          padding: EdgeInsets.all(MctSize.fifteen.getSize),
           reverse: true,
           shrinkWrap: true,
           itemCount: audioAndVideoListController.videoList.length,
@@ -47,7 +49,7 @@ class _MyPageLookState extends State<MyPageLook> {
             if (audioAndVideoListController.videoList.isNotEmpty) {
               print('${item.exerPath}');
               return Padding(
-                padding: const EdgeInsets.only(bottom: basicPadding),
+                padding: EdgeInsets.only(bottom: MctSize.fifteen.getSize),
                 child: InkWell(
                   onTap: () async {
                     var dir = (await getApplicationDocumentsDirectory()).path;
@@ -66,8 +68,8 @@ class _MyPageLookState extends State<MyPageLook> {
                       // mainAxisAlignment: MainAxisAlignment.start,
                       children: [
                         Padding(
-                          padding: const EdgeInsets.symmetric(
-                              horizontal: basicPadding),
+                          padding: EdgeInsets.symmetric(
+                              horizontal: MctSize.fifteen.getSize),
                           child: Column(
                             mainAxisAlignment: MainAxisAlignment.center,
                             crossAxisAlignment: CrossAxisAlignment.start,
@@ -76,15 +78,16 @@ class _MyPageLookState extends State<MyPageLook> {
                                 '${item.songTitle}',
                                 textAlign: TextAlign.left,
                                 style: TextStyle(
-                                    fontSize: textEightSize.sp,
-                                    color: white,
+                                    fontSize: MctSize.eighteen.getSize.sp,
+                                    color: MctColor.white.getMctColor,
                                     fontFamily: NOTO_MEDIUM),
                               ),
                               SizedBox(height: 2),
                               Text(
                                 audioAndVideoListController
                                     .convertDateFormat(item.exerTime),
-                                style: TextStyle(color: white),
+                                style: TextStyle(
+                                    color: MctColor.white.getMctColor),
                               ),
                             ],
                           ),
