+++ lib/widgets/play_jungganbo/game_timer_widget.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter/material.dart';
 import 'package:project_danso/common/color.dart';
-import 'package:project_danso/common/const.dart';
+import 'package:project_danso/common/contant.dart';
 
 class GameTimerWidget extends StatefulWidget {
   const GameTimerWidget({Key? key}) : super(key: key);
